requirements.txt
astroid==2.3.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.2
Flask==1.1.1
Flask-WTF==0.14.3
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
imageio==2.8.0
imutils==0.5.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
networkx==2.4
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
PyWavelets==1.1.1
requests==2.23.0
rsa==4.0
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
